feat(usersManager): profile user page was created
