+++ src/Controller/AbstractController.php
@@ -6,7 +6,6 @@ namespace App\Controller;
 
 use App\Router\Router;
 use App\Services\CsrfServiceProvider;
-use App\SuperGlobals\Get;
 use App\SuperGlobals\Session;
 use Exception;
 use Twig\Environment;
@@ -40,11 +39,9 @@ class AbstractController
             ));
         }
 
-        // on récupère l'utilisateur courant pour le rendre accessible globalement depuis twig
-        $parameters['app']['user'] = Session::get('user');
-        $parameters['app']['error'] = Get::get('error');
-        $parameters['app']['success'] = Get::get('success');
-
+        $parameters['app']['user'] = Session::get('user'); /* @phpstan-ignore-line */
+        $parameters['app']['error'] = Session::get('error'); /* @phpstan-ignore-line */
+        $parameters['app']['success'] = Session::get('success'); /** @phpstan-ignore-line */
         $twig = new Environment($loader, []);
         $template = $twig->load($view . '.html.twig');
         $template->display($parameters);
