+++ src/Dto/MyAccount/UpdateMyAccountIdentity.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\Dto\MyAccount;
 
-class  UpdateMyAccountIdentity
+class UpdateMyAccountIdentity
 {
     public function __construct(
         public string $id,
