--- templates/dashboard/admin/user.html.twig
@@ -1,150 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}DevBlog | Gestion des utilisateurs{% endblock %}
-
-{% block main_content %}
-    <div class="container mt-5">
-        <div class="row justify-content-between">
-
-            <div class="card">
-                <div class="card-header">
-                    <h4 class="card-header-title me-auto">
-                        Gestion des utilisateurs
-                    </h4>
-
-                    <ul class="nav nav-tabs nav-tabs-sm card-header-tabs">
-                        <li class="nav-item">
-                            <a href="/dashboard/users" class="nav-link active">Utilisateurs</a>
-                        </li>
-                        <li class="nav-item">
-                            <a href="#!" class="nav-link disabled">Articles</a>
-                        </li>
-                        <li class="nav-item">
-                            <a href="#!" class="nav-link disabled">Commentaires</a>
-                        </li>
-                        <li class="nav-item">
-                            <a href="#!" class="nav-link disabled">Contacts</a>
-                        </li>
-                    </ul>
-
-                </div>
-                <div class="card-body">
-                    <h1 class="display-4 text-center mb-3">
-                        Modifier le compte de {{ user.firstName }} {{ user.lastName }}
-                    </h1>
-                    {% if errors.global %}
-                        <div class="alert alert-danger" role="alert">
-                            {{ errors.global }}
-                        </div>
-                    {% endif %}
-
-                    <form action="/dashboard/user/{{ user.id }}/update" method="post" data-bitwarden-watching="1">
-                        <input type="hidden" name="_method" value="put">
-                        <div class="form-group mb-4">
-                            <label class="form-label">
-                                Prénom
-                            </label>
-                            <input type="text" name="firstName" id="firstName" class="form-control {% if errors.firstName %}is-invalid{% endif %}" value="{{ user.firstName }}" required>
-                            {% if errors.firstName %}
-                                <div id="firstName_error" class="invalid-feedback">
-                                    {{ errors.firstName }}
-                                </div>
-                            {% endif %}
-                        </div>
-                        <div class="form-group mb-4">
-                            <label class="form-label">
-                                Nom
-                            </label>
-                            <input type="text" name="lastName" id="lastName" class="form-control {% if errors.lastName %}is-invalid{% endif %}" value="{{ user.lastName }}" required>
-                            {% if errors.lastName %}
-                                <div id="lastName_error" class="invalid-feedback">
-                                    {{ errors.lastName }}
-                                </div>
-                            {% endif %}
-                        </div>
-
-                        <div class="form-group mb-4">
-                            <label class="form-label">
-                                Adresse email
-                            </label>
-                            <input type="email" name="email" id="email" class="form-control {% if errors.email %}is-invalid{% endif %}" value="{{ user.email }}" required>
-                            {% if errors.email %}
-                                <div id="email_error" class="invalid-feedback">
-                                    {{ errors.email }}
-                                </div>
-                            {% endif %}
-                        </div>
-
-                        <div class="form-check form-switch">
-                            <input type="hidden" value="off" name="isValidated">
-                            <input class="form-check-input" type="checkbox" role="switch" name="isValidated" id="isValidated" {% if user.isValidated == true %}checked{% endif %}>
-                            <label class="form-check-label" for="isAuthor">Compte validé ?</label>
-                        </div>
-                        <input type="hidden" name="id" value="{{ user.id }}">
-
-                        {% if csrf %}
-                            <input type="hidden" name="_csrf" value="{{ csrf }}">
-                        {% endif %}
-
-                        <input type="submit" class="btn btn-lg w-100 btn-primary my-3" value="Modifier">
-                    </form>
-                </div>
-            </div>
-
-            {# Modal de suppression #}
-            <div class="modal fade" tabindex="-1">
-                <div class="modal-dialog">
-                    <div class="modal-content">
-                        <div class="modal-header">
-                            <h5 class="modal-title">Supprimer un utilisateur</h5>
-                            <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
-                        </div>
-                        <div class="modal-body">
-                            <p>Supprimer le compte de <span class="modal__text__name"></span></p>
-                        </div>
-                        <div class="modal-footer">
-                            <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Annuler</button>
-                            <a type="button" class="btn btn-danger btn__user__delete">Supprimer</a>
-                        </div>
-                    </div>
-                </div>
-            </div>
-
-        </div>
-    </div>
-{% endblock %}
-
-{% block javascripts  %}
-    {{ parent() }}
-    <script>
-
-        // Modification et ouverture de la modal
-        modal = document.querySelector('.modal');
-        deleteButtons = document.querySelectorAll('.delete-button');
-        deleteButtons.forEach((e) => {
-            e.addEventListener('click', () => {
-                id = e.getAttribute('user-id')
-                firstName = e.getAttribute('user-first-name')
-                lastName = e.getAttribute('user-last-name')
-
-                deleteUserButton = document.querySelector('.btn__user__delete');
-                deleteUserButton.setAttribute('href', '/dashboard/user/' + id + '/delete')
-
-                text = modal.querySelector('.modal__text__name');
-                text.innerHTML = firstName + ' ' + lastName;
-                modal.style.display = 'block'
-                modal.classList.add('show');
-            })
-        })
-
-        // Fermuture de la modal
-        closeModal = document.querySelectorAll("[data-bs-dismiss=modal]");
-        closeModal.forEach((e) => {
-            e.addEventListener('click', () => {
-                console.log('On ferme la modal')
-                modal.style.display = 'none'
-                modal.classList.remove('show');
-            })
-        })
-    </script>
-{% endblock %}
